Add custom allocator support and related tests (foundation only)

+++ tests/API Tests/ConnectTests.cpp
@@ -377,7 +377,7 @@ TEST_SUITE("MqttClient Connect") {
     args.will = std::make_unique<kmMqtt::mqtt::Will>("will/topic");
     static constexpr uint8_t data[] = {1, 2, 3};
     args.will->payload =
-        std::make_unique<kmMqtt::mqtt::BinaryData>(sizeof(data), data);
+        std::make_unique<kmMqtt::mqtt::BinaryData>(static_cast<uint16_t>(sizeof(data)), data);
     args.will->willQos = kmMqtt::mqtt::Qos::QOS_1;
     args.will->retainWillMessage = true;
     testContext.tryConnect(kmMqtt::mqtt::ClientErrorCode::No_Error,
